# backend/swagger.yaml
openapi: 3.0.0
info:
  title: Dustid Contact Verification API
  version: 1.0.0
  description: Simple mock API for OTP verification and contact lookup.

servers:
  - url: https://dustid-backend-latest.onrender.com/
    description: Staging server
  - url: https://localhost:5000/
    description: Local development server (Port 5000)
  - url: https://localhost:3000/
    description: Local development server (Port 3000)

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Message:
      type: object
      properties:
        message:
          type: string
    Address:
      type: object
      properties:
        address1:
          type: string
        address2:
          type: string
        city:
          type: string
        province:
          type: string
        zip:
          type: string
        country:
          type: string
        first_name:
          type: string
        last_name:
          type: string
        phone:
          type: string
    Friend:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
        address:
          $ref: '#/components/schemas/Address'

paths:
  /:
    get:
      summary: Home route
      description: Basic check if API is running
      parameters:
        - name: embedded
          in: query
          required: false
          schema:
            type: string
          description: Shopify embedded app indicator
        - name: hmac
          in: query
          required: false
          schema:
            type: string
          description: Shopify HMAC signature
        - name: host
          in: query
          required: false
          schema:
            type: string
          description: Shopify embedded app host value
        - name: id_token
          in: query
          required: false
          schema:
            type: string
          description: Shopify ID token for embedded apps
        - name: locale
          in: query
          required: false
          schema:
            type: string
          description: Locale sent by Shopify
        - name: session
          in: query
          required: false
          schema:
            type: string
          description: Shopify session token identifier
        - name: shop
          in: query
          required: false
          schema:
            type: string
          description: Shopify shop domain
        - name: timestamp
          in: query
          required: false
          schema:
            type: string
          description: Shopify timestamp
      responses:
        '200':
          description: Welcome page
          content:
            text/html:
              schema:
                type: string

  /api/auth:
    get:
      summary: Shopify auth start
      parameters:
        - name: shop
          in: query
          required: true
          schema:
            type: string
        - name: embedded
          in: query
          required: false
          schema:
            type: string
      responses:
        '302':
          description: Redirect to Shopify authorization

  /verify:
    post:
      summary: Generate OTP
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - phoneNumber
              properties:
                phoneNumber:
                  type: string
                  example: "254712345678"
      responses:
        '200':
          description: OTP sent successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
        '400':
          description: Missing phone number
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
        '404':
          description: Phone number not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'

  /validate-otp:
    post:
      summary: Validate OTP
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - phoneNumber
                - otp
              properties:
                phoneNumber:
                  type: string
                  example: "254712345678"
                otp:
                  type: string
                  example: "123456"
      responses:
        '200':
          description: OTP validated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: OTP validated successfully.
                  token:
                    type: string
                    example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        '400':
          description: Missing phone number or OTP
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
        '401':
          description: Invalid OTP
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'

  /friends:
    get:
      summary: Get all friends
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of friends
          content:
            application/json:
              schema:
                type: object
                properties:
                  friends:
                    type: array
                    items:
                      $ref: '#/components/schemas/Friend'
        '401':
          description: Missing or invalid token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
        '404':
          description: Contact not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'

  /friends/{friendId}:
    get:
      summary: Get a friend's full details including address
      security:
        - bearerAuth: []
      parameters:
        - name: friendId
          in: path
          required: true
          schema:
            type: string
          example: friend101
      responses:
        '200':
          description: Friend object
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Friend'
        '401':
          description: Missing or invalid token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
        '403':
          description: Friend not in your list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
        '404':
          description: Contact not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'

  /search:
    post:
      summary: Search friends
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - phoneNumber
              properties:
                phoneNumber:
                  type: string
                  example: "254712345678"
                query:
                  type: string
                  description: Optional search term
                  example: "bob"
      responses:
        '200':
          description: List of friends
          content:
            application/json:
              schema:
                type: object
                properties:
                  results:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                        name:
                          type: string
                        email:
                          type: string
        '400':
          description: Missing phone number
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
        '401':
          description: Missing or invalid token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
        '404':
          description: Contact not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'

  /pre-fill-checkout:
    post:
      summary: Pre-fill checkout with selected contact
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - contact
              properties:
                contact:
                  $ref: '#/components/schemas/Friend'
      responses:
        '200':
          description: Contact received
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  contact:
                    $ref: '#/components/schemas/Friend'

  /profile:
    post:
      summary: Get contact profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - phoneNumber
              properties:
                phoneNumber:
                  type: string
                  example: "254712345678"
      responses:
        '200':
          description: Contact object
          content:
            application/json:
              schema:
                type: object
                properties:
                  contact:
                    type: object
                    properties:
                      id:
                        type: string
                      phoneNumber:
                        type: string
                      name:
                        type: string
                      friends:
                        type: array
                        items:
                          $ref: '#/components/schemas/Friend'
        '400':
          description: Missing phone number
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
        '404':
          description: Contact not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
  /api/draft-order:
    post:
      summary: Create a draft order for gift checkout
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - shop
                - items
                - contact
              properties:
                shop:
                  type: string
                  example: "dustid-shop.myshopify.com"
                items:
                  type: array
                  items:
                    type: object
                    properties:
                      variant_id:
                        type: integer
                      quantity:
                        type: integer
                contact:
                  $ref: '#/components/schemas/Friend'
      responses:
        '200':
          description: Shopify checkout URL created successfully.
          content:
            application/json:
              schema:
                type: object
                properties:
                  invoice_url:
                    type: string
                  message:
                    type: string
        '400':
          description: Unable to build Shopify checkout URL from request payload.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'